requirements.txt
pydantic==2.6.0
openai>=1.51.0
httpx>=0.27.0
pyyaml==6.0.1
python-dotenv